backend/requirements.txt
Django==4.2.24
djangorestframework==3.14.0
django-filter==23.3
psycopg[binary]==3.1.18
python-decouple==3.8
dj-database-url==2.1.0
django-cors-headers==4.3.1
Pillow==10.0.1
pytest-django==4.8.0
requests==2.31.0